requirements.txt
streamlit==1.3.0
pandas==1.3.4
numpy==1.21.4
matplotlib==3.5.1
altair==4.1.0
plotly-express==0.4.1
openpyxl==3.0.9